persona/emergent_misalignment/requirements.txt
torch
transformers
datasets
accelerate
peft
bitsandbytes
fire
scipy
numpy
pandas
tqdm
wandb
scikit-learn
matplotlib
seaborn
jupyter
ipykernel
pyyaml
jsonlines
openai
anthropic
tiktoken
vllm